backend/requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
mangum>=0.19.0
boto3>=1.35.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
